requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-bootstrap==0.2.4
django-filter==2.4.0
django-google-charts==0.1.1
django-on-heroku==1.1.2
gunicorn==20.1.0
idna==2.10
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
psycopg2-binary==2.8.6
PyQt5==5.15.4
PyQt5-sip==12.8.1
pyTelegramBotAPI==3.7.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
